solved f to c and changed int to floar

+++ Jan16/Exercises/small/c_to_f.py
@@ -1,4 +1,4 @@
-cel = int(input("Type celsius: "))
+cel = float(input("Type celsius: "))
 
 
 def conv(cel):
